# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * IFA Lite API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Broker {
  id: number;
  ifaRef?: string;
  brokerNo?: string;
  fimbraNo?: string;
  brokerName: string;
  tradingName?: string;
  addressLine1?: string;
  addressLine2?: string;
  town?: string;
  county?: string;
  postcode?: string;
  telephone?: string;
  fax?: string;
  email?: string;
  initials?: string;
  dateChecked?: string;
  fcaReference?: string;
  annuityToba?: boolean;
  status?: string;
  sentDate?: string;
  grade?: string;
  nextDiaryDate?: string;
  ifaMemberNo?: string;
  brokerManager?: string;
  keyAccount?: string;
  partnerCode?: string;
  region?: string;
  createdBy?: string;
  createdDate?: string;
  amendedBy?: string;
  amendedDate?: string;
}

export interface CreateBroker {
  ifaRef?: string;
  brokerNo?: string;
  fimbraNo?: string;
  brokerName: string;
  tradingName?: string;
  addressLine1?: string;
  addressLine2?: string;
  town?: string;
  county?: string;
  postcode?: string;
  telephone?: string;
  fax?: string;
  email?: string;
  initials?: string;
  dateChecked?: string;
  fcaReference?: string;
  annuityToba?: boolean;
  status?: string;
  sentDate?: string;
  grade?: string;
  nextDiaryDate?: string;
  ifaMemberNo?: string;
  brokerManager?: string;
  keyAccount?: string;
  partnerCode?: string;
  region?: string;
}

export interface UpdateBroker {
  ifaRef?: string;
  brokerNo?: string;
  fimbraNo?: string;
  brokerName?: string;
  tradingName?: string;
  addressLine1?: string;
  addressLine2?: string;
  town?: string;
  county?: string;
  postcode?: string;
  telephone?: string;
  fax?: string;
  email?: string;
  initials?: string;
  dateChecked?: string;
  fcaReference?: string;
  annuityToba?: boolean;
  status?: string;
  sentDate?: string;
  grade?: string;
  nextDiaryDate?: string;
  ifaMemberNo?: string;
  brokerManager?: string;
  keyAccount?: string;
  partnerCode?: string;
  region?: string;
}

export interface Contact {
  id: number;
  brokerId: number;
  reference?: string;
  title?: string;
  initials?: string;
  forename?: string;
  surname?: string;
  salutation?: string;
  position?: string;
  addressLine1?: string;
  addressLine2?: string;
  addressLine3?: string;
  addressLine4?: string;
  homeTelephone?: string;
  mobileTelephone?: string;
  emailAddress?: string;
  paidByBacs?: boolean;
  bankSortCode?: string;
  bankAccountNo?: string;
  bankAccountName?: string;
  bankReference?: string;
  useNetworkTerms?: boolean;
  defaultAdviceType?: string;
  defaultRemunerationBasis?: string;
  defaultDistributionChannel?: string;
  restrictedAdvice?: string;
  simplifiedAdvice?: string;
  nonAdvised?: string;
  networkIfa?: string;
  networkName?: string;
  networkPostcode?: string;
  tiedAgent?: boolean;
  principalAgentRef?: string;
  isPrincipal?: boolean;
  network?: boolean;
}

export interface CreateContact {
  reference?: string;
  title?: string;
  initials?: string;
  forename?: string;
  surname?: string;
  salutation?: string;
  position?: string;
  addressLine1?: string;
  addressLine2?: string;
  addressLine3?: string;
  addressLine4?: string;
  homeTelephone?: string;
  mobileTelephone?: string;
  emailAddress?: string;
  paidByBacs?: boolean;
  bankSortCode?: string;
  bankAccountNo?: string;
  bankAccountName?: string;
  bankReference?: string;
  useNetworkTerms?: boolean;
  defaultAdviceType?: string;
  defaultRemunerationBasis?: string;
  defaultDistributionChannel?: string;
  restrictedAdvice?: string;
  simplifiedAdvice?: string;
  nonAdvised?: string;
  networkIfa?: string;
  networkName?: string;
  networkPostcode?: string;
  tiedAgent?: boolean;
  principalAgentRef?: string;
  isPrincipal?: boolean;
  network?: boolean;
}

export interface UpdateContact {
  reference?: string;
  title?: string;
  initials?: string;
  forename?: string;
  surname?: string;
  salutation?: string;
  position?: string;
  addressLine1?: string;
  addressLine2?: string;
  addressLine3?: string;
  addressLine4?: string;
  homeTelephone?: string;
  mobileTelephone?: string;
  emailAddress?: string;
  paidByBacs?: boolean;
  bankSortCode?: string;
  bankAccountNo?: string;
  bankAccountName?: string;
  bankReference?: string;
  useNetworkTerms?: boolean;
  defaultAdviceType?: string;
  defaultRemunerationBasis?: string;
  defaultDistributionChannel?: string;
  restrictedAdvice?: string;
  simplifiedAdvice?: string;
  nonAdvised?: string;
  networkIfa?: string;
  networkName?: string;
  networkPostcode?: string;
  tiedAgent?: boolean;
  principalAgentRef?: string;
  isPrincipal?: boolean;
  network?: boolean;
}

export interface Note {
  id: number;
  brokerId: number;
  noteType?: string;
  description?: string;
  oldValue?: string;
  newValue?: string;
  updatedBy?: string;
  updatedDate?: string;
}

export interface CreateNote {
  noteType?: string;
  description?: string;
  oldValue?: string;
  newValue?: string;
  updatedBy?: string;
  updatedDate?: string;
}

export interface RetirementIncome {
  id?: number;
  brokerId?: number;
  npaAdviserCharges?: number;
  npaAmount?: number;
  npaCommission?: number;
  npaExpenseDiscount?: number;
  npaMarketingAllowance?: number;
  pipaAdviserCharges?: number;
  pipaAmount?: number;
  pipaCommission?: number;
  pipaExpenseDiscount?: number;
  pipaMarketingAllowance?: number;
  prpAdviserCharges?: number;
  prpAmount?: number;
  prpCommission?: number;
  prpExpenseDiscount?: number;
  prpMarketingAllowance?: number;
}

export interface EquityRelease {
  id?: number;
  brokerId?: number;
  mortgagePermissions?: boolean;
  erlmToba?: boolean;
  flexibleBrokerRate?: number;
  flexibleMinimumAmount?: number;
  flexibleNetworkRate?: number;
  flexibleTrailCommission?: boolean;
  lumpSumBrokerRate?: number;
  lumpSumMinimumAmount?: number;
  lumpSumNetworkRate?: number;
  packagingFee?: number;
  applicationFee?: number;
  ltvPercent?: number;
  lumpSumPackagingFee?: number;
  lumpSumApplicationFee?: number;
  lumpSumLtvPercent?: number;
}

export type ListBrokersParams = {
  postcode?: string;
  ifaReference?: string;
  ifaName?: string;
  authorised?: boolean;
  cancelled?: boolean;
  duplicateRecord?: boolean;
  revoked?: boolean;
};